fix(web): render Navbar compound components from a client boundary
ratio-ui 2.8 -> 2.14 made Navbar a 'use client' module, so its compound
statics (Navbar.Brand, Navbar.Content) only exist inside the client
bundle. SiteNavbar — a server component — got undefined from the
client-reference proxy, crashing every page that renders the navbar
('Element type is invalid ... got: undefined', staging 500 since
2026-07-21, CD frozen).

Split SiteNavbar: the server half keeps fetching site settings and
translations, and a new SiteNavbarClient owns the Navbar compound
rendering. Verified locally: frontpage went 500 -> 200; production
build green. Footer/List compound usage in server components is fine —
those modules ship without the client directive.
